Getting the Save Button to show in the actionBar
This took some effort. A lot of back and forth
between Perplexity and me. The overall response
from Perplexity was trash, pure trash where even
repeated questions and pointing out issues couldn't
make the platform to respond correctly.

Unless, I looked through the HTML and got the structure
of the buttons which I then shared with Perplexity and
got the final working code.